html_to_markdown: Move `TableHandler` out of `rustdoc`
