+++ crates/html_to_markdown/src/html_to_markdown.rs
@@ -16,7 +16,9 @@ use html5ever::tendril::TendrilSink;
 use html5ever::tree_builder::TreeBuilderOpts;
 use markup5ever_rcdom::RcDom;
 
-use crate::markdown::{HeadingHandler, ListHandler, ParagraphHandler, StyledTextHandler};
+use crate::markdown::{
+    HeadingHandler, ListHandler, ParagraphHandler, StyledTextHandler, TableHandler,
+};
 use crate::markdown_writer::{HandleTag, MarkdownWriter};
 
 /// Converts the provided HTML to Markdown.
@@ -27,11 +29,11 @@ pub fn convert_html_to_markdown(html: impl Read) -> Result<String> {
         Box::new(ParagraphHandler),
         Box::new(HeadingHandler),
         Box::new(ListHandler),
+        Box::new(TableHandler::new()),
         Box::new(StyledTextHandler),
         Box::new(structure::rustdoc::RustdocChromeRemover),
         Box::new(structure::rustdoc::RustdocHeadingHandler),
         Box::new(structure::rustdoc::RustdocCodeHandler),
-        Box::new(structure::rustdoc::RustdocTableHandler::new()),
         Box::new(structure::rustdoc::RustdocItemHandler),
     ];
 
@@ -51,11 +53,11 @@ pub fn convert_rustdoc_to_markdown(html: impl Read) -> Result<String> {
         Box::new(ParagraphHandler),
         Box::new(HeadingHandler),
         Box::new(ListHandler),
+        Box::new(TableHandler::new()),
         Box::new(StyledTextHandler),
         Box::new(structure::rustdoc::RustdocChromeRemover),
         Box::new(structure::rustdoc::RustdocHeadingHandler),
         Box::new(structure::rustdoc::RustdocCodeHandler),
-        Box::new(structure::rustdoc::RustdocTableHandler::new()),
         Box::new(structure::rustdoc::RustdocItemHandler),
     ];
